Working VE and IGN table comms

+++ table.h
@@ -1,13 +1,13 @@
 /*
 This file is used for everything related to maps/tables including their definition, functions etc
 */
-
+#include <Arduino.h>
 struct table {
   //All tables must be the same size for simplicity
   const static int xSize = 8;
   const static int ySize = 8;
   
-  int values[ySize][xSize];
+  byte values[ySize][xSize];
   int axisX[xSize];
   int axisY[ySize];
